+++ Source/Vcl/Markdown4D.Vcl.Viewer.pas
@@ -197,6 +197,7 @@ uses
   Vcl.Imaging.jpeg,
   Vcl.Imaging.GIFImg,
   Markdown4D.Image.Svg.Native,
+  Markdown4D.Vcl.ImageDecoder,
   Markdown4D.Ast.Interfaces,
   Markdown4D.Layout.Defaults,
   Markdown4D.Layout.HitTest,